Refactor surah sidebar helpers
